chore(genie): move tsconfig project registry
