+++ tsconfig.check.json.genie.ts
@@ -1,6 +1,6 @@
 import type { TSConfigArgs } from './packages/@overeng/genie/src/runtime/mod.ts'
 import { tsconfigJson } from './packages/@overeng/genie/src/runtime/mod.ts'
-import { rootTsconfigProjects } from './tsconfig.projects.ts'
+import { rootTsconfigProjects } from './genie/tsconfig-projects.ts'
 
 // This file is meant for convenience to built all TS projects in the workspace at once
 export default tsconfigJson({
