+++ tsconfig.emit.json.genie.ts
@@ -1,6 +1,6 @@
 import { isTsconfigReferenceTarget } from './packages/@overeng/genie/src/runtime/composition/mod.ts'
 import { tsconfigJson, type TSConfigArgs } from './packages/@overeng/genie/src/runtime/mod.ts'
-import { rootTsconfigProjects } from './tsconfig.projects.ts'
+import { rootTsconfigProjects } from './genie/tsconfig-projects.ts'
 
 export default tsconfigJson({
   files: [],
